+++ api/service/notification-type.service.ts
@@ -2,10 +2,11 @@ import NotificationTypeModel from "../model/notification-type.model";
 import { z } from "zod";
 import AppError from "../utils/app-error";
 import { StatusCodes } from "http-status-codes";
+import NotificationSettingsModel from "../model/notification-settings.model";
 
 export const getNotificationTypesService = async () => {
   try {
-    return await NotificationTypeModel.find().sort("-createdAt");
+    return await NotificationTypeModel.find().sort("-name");
   } catch (error) {
     throw error;
   }
@@ -30,13 +31,28 @@ export const createNotificationTypeService = async (data: {
     if (existingData)
       throw new AppError(
         `Notification type with the name ${validData.name} already exists. Please use another name.`,
-        StatusCodes.BAD_REQUEST,
+        StatusCodes.BAD_REQUEST
       );
 
-    return await NotificationTypeModel.create({
+    const type = await NotificationTypeModel.create({
       name: validData.name,
       description: validData.description,
     });
+
+    // update settings for all users
+    await NotificationSettingsModel.updateMany(
+      {},
+      {
+        $push: {
+          preferences: {
+            type: type._id,
+            enabled: true,
+          },
+        },
+      }
+    );
+
+    return type;
   } catch (error) {
     throw error;
   }